+++ packages/PowerPack/rspack.config.js
@@ -3,6 +3,7 @@ const {bundler} = require('@datagrok/build-config');
 module.exports = bundler({
   "externals": {
     "datagrok-api/u2core": "DG.U2",
+    "codemirror": false,
     "exceljs": false
   }
 });